requirements.txt
aioboto3==13.4.0
aiobotocore==2.18.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aioitertools==0.12.0
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asgi-lifespan==2.1.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==25.1.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
boto3==1.36.1
botocore==1.36.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
coverage==7.4.1
cryptography==44.0.1
dataclasses-json==0.6.7
deepdiff==8.2.0
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
emoji==2.14.1
exceptiongroup==1.2.2
faiss-cpu==1.10.0
Faker==22.6.0
fastapi==0.115.8
filetype==1.2.0
frozenlist==1.5.0
gotrue==1.3.1
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-openai==0.3.7
langchain-pinecone==0.3.1
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.3.11
lxml==5.3.1
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.1.0
mypy-extensions==1.0.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
openai==1.65.2
orderly-set==5.3.0
orjson==3.10.15
packaging==24.2
passlib==1.7.4
pinecone==6.0.1
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
postgrest==0.11.0
propcache==0.3.0
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
pypdf==5.3.1
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dateutil==2.9.0
python-dotenv==1.0.1
python-iso639==2025.2.18
python-jose==3.4.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
PyYAML==6.0.2
RapidFuzz==3.12.2
realtime==1.0.6
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.45.3
storage3==0.6.1
StrEnum==0.4.15
supabase==1.2.0
supafunc==0.2.3
tenacity==9.0.0
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
unstructured==0.16.23
unstructured-client==0.22.0
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
websockets==12.0
wrapt==1.17.2
yarl==1.18.3
zstandard==0.23.0
